Apply Rider's Full Cleanup to the merged work
eng/verify.ps1 runs jb cleanupcode with the Full Cleanup profile and fails on any
diff. Nothing had reached that gate in days because the Prettier check ahead of it
was failing first, so the drift accumulated across the issue handoff and predates
part of it. Layout only: field ordering, a blank line after guard clauses, and one
redundant cast.

+++ src/WSGM/Core/DisplayScale.cs
@@ -57,7 +57,7 @@ public static class DisplayScale
                 continue;
             }
 
-            captured.Add(new DisplayScaleEntry { DeviceName = name, Percent = (int)current });
+            captured.Add(new DisplayScaleEntry { DeviceName = name, Percent = current });
             // A surviving snapshot means the previous game-mode session did not
             // finish restoring every display.  A dock/undock can expose a different
             // active source before recovery runs; never force that new display to
